Rigid bodies no longer store an ID; objects handle this as they should. Also properly set some pointers to NULL for scenes / islands. While it isn't strictly necessary, it'll help prevent loops leaking into the wrong linked lists.

+++ src/physicsShared.h
@@ -4,6 +4,6 @@
 #include "physicsSettings.h"
 #include <stdint.h>
 
-typedef uint_least16_t physicsBodyIndex_t;
+typedef uint_least8_t physicsBodyIndex_t;
 
 #endif
